+++ packages/pluggableWidgets/feedback-native/typings/FeedbackProps.d.ts
@@ -12,6 +12,17 @@ export interface FeedbackProps<Style> {
     sprintrapp: string;
     allowScreenshot: boolean;
     logo?: DynamicValue<NativeImage>;
+    titleSendFeedback?: DynamicValue<string>;
+    titleSending?: DynamicValue<string>;
+    titleResult?: DynamicValue<string>;
+    labelFeedbackInput?: DynamicValue<string>;
+    labelIncludeScreenshot?: DynamicValue<string>;
+    buttonCancel?: DynamicValue<string>;
+    buttonSend?: DynamicValue<string>;
+    buttonOk?: DynamicValue<string>;
+    accessibilityLabelFeedbackButton?: DynamicValue<string>;
+    messageSuccess?: DynamicValue<string>;
+    messageError?: DynamicValue<string>;
 }
 
 export interface FeedbackPreviewProps {
@@ -28,4 +39,15 @@ export interface FeedbackPreviewProps {
     sprintrapp: string;
     allowScreenshot: boolean;
     logo: { type: "static"; imageUrl: string } | { type: "dynamic"; entity: string } | null;
+    titleSendFeedback: string;
+    titleSending: string;
+    titleResult: string;
+    labelFeedbackInput: string;
+    labelIncludeScreenshot: string;
+    buttonCancel: string;
+    buttonSend: string;
+    buttonOk: string;
+    accessibilityLabelFeedbackButton: string;
+    messageSuccess: string;
+    messageError: string;
 }