+++ packages/pluggableWidgets/signature-native/src/__tests__/Signature.android.spec.tsx
@@ -22,13 +22,14 @@ jest.mock("react-native", () => {
 jest.mock("react-native/Libraries/Utilities/Platform", () => {
     const Platform = jest.requireActual("react-native/Libraries/Utilities/Platform");
     Platform.OS = "android";
+    Platform.default = { ...Platform.default, OS: "android" };
     return Platform;
 });
 
 jest.mock("react-native-webview", () => {
-    const { View } = require("react-native");
+    const { View } = jest.requireActual("react-native");
 
-    const WebView = (props: any) => {
+    const WebView = (props: any): ReactElement => {
         return <View {...props} testID="mockWebView" />;
     };
 
@@ -44,7 +45,11 @@ jest.mock("react-native/Libraries/Components/Touchable/TouchableNativeFeedback",
     TouchableNativeFeedback.SelectableBackgroundBorderless = jest.fn(() => "SelectableBackgroundBorderless");
     TouchableNativeFeedback.Ripple = jest.fn(() => "Ripple");
     TouchableNativeFeedback.canUseNativeForeground = jest.fn(() => true);
-    return TouchableNativeFeedback;
+
+    return {
+        __esModule: true,
+        default: TouchableNativeFeedback
+    };
 });
 
 const mockImageSource: any = {