+++ packages/pluggableWidgets/listview-swipe-native/CHANGELOG.md
@@ -6,9 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   Updated react-native-gesture-handler from version v2.24.0 to v2.30.0.
+
 ## [3.3.0] - 2025-3-31
 
--  Updated react-native-gesture-handler from version v2.16.2 to 2.24.0
+-   Updated react-native-gesture-handler from version v2.16.2 to 2.24.0
 
 ## [3.2.0] - 2024-12-3