+++ packages/pluggableWidgets/signature-native/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   Renamed `onSave` action to `onSignEnd` to match web signature widget naming convention
 -   Removed `imageAttribute` property in favor of unified `imageSource` property
+-   Updated react-native-webview from version v13.13.2 to v13.16.1.
 
 ## [2.3.0] - 2025-7-7